requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
docx2txt==0.6
EbookLib==0.15
extract-msg==0.23.1
future==0.17.1
idna==2.8
ImageHash==4.0
IMAPClient==2.1.0
lxml==4.4.0
numpy==1.17.0
olefile==0.46
pdfminer.six==20181108
Pillow==6.1.0
pkg-resources==0.0.0
pocketsphinx==0.1.3
pycryptodome==3.8.2
PyPDF2==1.26.0
pytesseract==0.2.7
python-dateutil==2.8.0
python-docx==0.8.10
python-pptx==0.6.5
pytz==2019.2
PyWavelets==1.0.3
regex==2019.6.8
requests==2.22.0
scipy==1.3.1
six==1.10.0
sortedcontainers==2.1.0
soupsieve==1.9.2
SpeechRecognition==3.6.3
textract==1.6.1
tzlocal==1.5.1
urllib3==1.25.3
xlrd==1.0.0
XlsxWriter==1.1.8
zenpy==2.0.12
